+++ benchmarks/profile_dram.py
@@ -10,8 +10,8 @@ dotpath, _ = IMPLEMENTATIONS[impl_name]
 fn = load_fn(dotpath)
 
 M = N = K = sz
-A = torch.randn(M, K, dtype=torch.bfloat16, device='cuda')
-B = torch.randn(K, N, dtype=torch.bfloat16, device='cuda')
+A = torch.randn(M, K, dtype=torch.float32, device='cuda')
+B = torch.randn(K, N, dtype=torch.float32, device='cuda')
 
 for _ in range(3):
     fn(A, B)
