+++ benchmarks/profile_s3.py
@@ -1,10 +1,10 @@
 """Profile best Stage 3 kernel (tm8_tn8_bm128_bn128_bk32_u8) with ncu."""
 import torch
-from mymatmul.gpu.matmul_cuda_s3 import matmul_s3_tm8_tn8_bm128_bn128_bk32_u8 as matmul_s3_best
+from mymatmul.gpu.cuda_core.matmul_cuda_s3 import matmul_s3_tm8_tn8_bm128_bn128_bk32_u8 as matmul_s3_best
 
 M, K, N = 4096, 4096, 4096
-A = torch.randn(M, K, dtype=torch.bfloat16, device="cuda")
-B = torch.randn(K, N, dtype=torch.bfloat16, device="cuda")
+A = torch.randn(M, K, dtype=torch.float32, device="cuda")
+B = torch.randn(K, N, dtype=torch.float32, device="cuda")
 
 # Warmup
 for _ in range(3):
